requirements.txt
Django==5.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
django-guardian==2.4.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
django-xhtml2pdf==0.0.3
xlsxwriter==3.1.9
whitenoise==6.6.0
Pillow==10.1.0
python-decouple==3.8
